fix: accounts sync on switching
