+++ src/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './local'
 export * from './types'
 export * from './utils'
+export * from './chains'
